api/requirements.txt
fastapi
uvicorn[standard]
watchdog
pydantic-settings
chromadb
ollama
pypdf
python-docx
mammoth
python-multipart
tqdm
